WINE_QUALITY_PREDICTION_END_TO_END_DEPLOYMENT/requirements.txt
pandas
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
Flask
Flask-Cors
